refactor(overlay): extract WS and keyboard handlers from index.ts
Extract ~590 lines from index.ts (1642→1052 lines) into:
- ws-handler.ts (415 lines) — panel→overlay WS message handling
- keyboard-handler.ts (266 lines) — delete/copy/paste/cut/duplicate

Both use dependency injection for index.ts-local values and import
shared modules directly. No behavior changes.
All 946 tests pass, esbuild bundle builds clean.
